+++ packages/toolbars/web/src/Toolbar/ButtonComponents/ColorPickerButton.jsx
@@ -94,7 +94,7 @@ class ColorPickerButton extends Component {
                     {renderUserColors()}
                   </div>
                   <hr className={mergedStyles.colorPicker_separator} />
-                  <div className={mergedStyles.colorPicker_buttons_container}>
+                  <div className={mergedStyles.colorPicker_bottom_container}>
                     {renderResetColorButton()}
                     {renderAddColorButton()}
                   </div>