+++ CHANGELOG.md
@@ -9,6 +9,13 @@
 > - :house: Internal
 > - :nail_care: Polish
 
+## [Unreleased]
+### :house: Internal
+- `unsupported-blocks`
+  - [#1987](https://github.com/wix/ricos/pull/1987) unsupported-blocks test
+- `color-picker`
+  - [#2000](https://github.com/wix/ricos/pull/2000) remove position fixed from reset button
+
 ## [Unreleased]  
 
 ### :bug: Bug Fix
@@ -22,6 +29,8 @@
   </summary>
 
 ### :house: Internal
+- `e2e`
+  - [#1983](https://github.com/wix/ricos/commit/1983) extend url limit length
 - `editor`
   - [#1996](https://github.com/wix/ricos/commit/1996) add support for new error codes, split `ErrorToast` component for scalability
 - `editor-common`